cleanup from restructure
