+++ frontend/tsconfig.json
@@ -10,7 +10,6 @@
       "@context/*": ["src/context/*"],
       "@pages/*": ["src/pages/*"],
       "@styles/*":["src/styles/*"],
-      "@types/*": ["src/types/*"],
   }
 },
   "extends": "expo/tsconfig.base"
